feat fix toàn bộ trang home lại
- thay đổi layout:
+ bỏ sản phẩm nổi bật
+ sửa layout new và sale
+ bỏ sản phẩm chó mèo thay bằng dịch vụ
+ thêm tiêu đề cho dịch vụ và bài viết
+ tách 1 số component
+ sửa seach
