+++ FE/src/App.tsx
@@ -22,7 +22,6 @@ import CategoryList from "./admin/category/category";
 import OrderList from "./admin/order/order";
 import UserList from "./admin/user/user";
 import ServiceList from "./admin/service/service";
-// import SystemSettings from "./admin/setting/setting";
 import EmployeeList from "./admin/employee/employee";
 import Payment from "./pages/payment/payment";
 import AboutUs from "./pages/about-us/about-us";
