+++ FE/src/components/layout/PageLayout.tsx
@@ -3,10 +3,10 @@ import { SearchProvider } from "../searchContext";
 
 import {} from "react";
 
-import Header from "../header";
+import Header from "./header";
 import { Outlet } from "react-router-dom";
-import Footer from "../footer";
-import Navigation from "../navigation";
+import Footer from "./footer";
+import Navigation from "./navigation";
 import ScrollToTop from "../ScrollToTop";
 function PageLayout() {
   return (
