+++ FE/src/pages/home/home.tsx
@@ -4,12 +4,10 @@ import { FaUserEdit, FaCalendarAlt } from "react-icons/fa";
 import { Button, Space } from "antd";
 import Slider from "react-slick";
 import { useState, useEffect, useRef } from "react";
-import SaleProduct from "../../components/saleproduct";
-import HotProduct from "../../components/hotproduct";
-import NewProduct from "../../components/newproduct";
+import ProductSlider from "../../components/home/ProductSlider";
 import CateProduct from "../../components/cateproduct";
-import "slick-carousel/slick/slick.css"; // Import CSS cho slick
-import "slick-carousel/slick/slick-theme.css"; // Import theme CSS
+import "slick-carousel/slick/slick.css";
+import "slick-carousel/slick/slick-theme.css";
 import ENV_VARS from "../../../config";
 import productsApi from "../../api/productsApi";
 import categoryApi from "../../api/categoryApi";
@@ -22,12 +20,8 @@ export default function Home() {
   const [saleProduct, setSaleProduct] = useState([]);
   const [hotProduct, setHotProduct] = useState([]);
   const [blogs, setBlogs] = useState<any[]>([]);
-  const [productsByCategory, setProductsByCategory] = useState<{
-    [key: string]: any[];
-  }>({}); // Lưu sản phẩm theo danh mục
-  const [categories, setCategories] = useState<{ _id: string; name: string }[]>(
-    []
-  );
+  const [productsByCategory, setProductsByCategory] = useState<{ [key: string]: any[] }>({});
+  const [categories, setCategories] = useState<{ _id: string; name: string }[]>([]);
 
   const images = [
     "/images/banners/1.png",
@@ -37,12 +31,11 @@ export default function Home() {
     "/images/banners/5.png",
   ];
 
-  const sliderRef = useRef<any>(null); // Ref cho Slider
+  const sliderRef = useRef<any>(null);
 
   useEffect(() => {
     const fetchProducts = async () => {
       try {
-        // Lấy danh mục
         const categoriesResponse = await categoryApi.getCategoriesActive();
         const categoriesData = await categoriesResponse.data.result;
         setCategories(categoriesData);
@@ -70,35 +63,29 @@ export default function Home() {
     fetchProducts();
   }, []);
 
-  // Lấy sản phẩm theo danh mục sau khi categories được cập nhật
   useEffect(() => {
     const fetchProductsByCategory = async () => {
-      if (categories.length === 0) return; // Không làm gì nếu categories rỗng
+      if (categories.length === 0) return;
 
       try {
         const categoryPromises = categories.map(async (category) => {
-          const productResponse = await productsApi.getProductByCategoryID(
-            category._id
-          );
+          const productResponse = await productsApi.getProductByCategoryID(category._id);
           const productData = await productResponse.data.result;
           const limitedProducts = productData ? productData.slice(0, 8) : [];
           return { [category.name]: limitedProducts };
         });
 
         const categoryProducts = await Promise.all(categoryPromises);
-        const productsMap = categoryProducts.reduce((acc, curr) => {
-          return { ...acc, ...curr };
-        }, {});
+        const productsMap = categoryProducts.reduce((acc, curr) => ({ ...acc, ...curr }), {});
         setProductsByCategory(productsMap);
       } catch (error) {
         console.error("Error fetching products by category:", error);
-        setProductsByCategory({}); // Reset nếu lỗi
+        setProductsByCategory({});
       }
     };
     fetchProductsByCategory();
   }, [categories]);
 
-  // Cấu hình settings cho Slider
   const settings = {
     dots: true,
     infinite: true,
@@ -109,14 +96,13 @@ export default function Home() {
     autoplaySpeed: 2000,
     arrows: false,
     fade: true,
-    // Tùy chỉnh style cho dots
     appendDots: (dots) => (
-      <div className="custom-dots-container flex justify-center py-4">
+      <div className="flex justify-center py-4 custom-dots-container">
         <ul>{dots}</ul>
       </div>
     ),
     customPaging: () => (
-      <div className="w-3 h-3 rounded-full bg-white transition-all duration-300"></div>
+      <div className="w-3 h-3 transition-all duration-300 bg-white rounded-full"></div>
     ),
   };
 
@@ -130,7 +116,7 @@ export default function Home() {
               <img
                 src={image}
                 alt={`Banner ${index + 1}`}
-                className="w-full object-cover"
+                className="object-cover w-full"
               />
             </div>
           ))}
@@ -139,72 +125,68 @@ export default function Home() {
 
       {/* Sản phẩm mới */}
       <div className="relative mt-[30px] rounded-lg p-6 px-4 sm:px-[40px] lg:px-[154px]">
-        <NewProduct data={newProduct} />
+        <ProductSlider title="SẢN PHẨM MỚI" data={newProduct} />
       </div>
 
       {/* Sản phẩm giảm giá */}
       <div className="relative mt-[30px] rounded-lg p-6 px-4 sm:px-[40px] lg:px-[154px]">
-        <SaleProduct data={saleProduct} />
+        <ProductSlider title="SẢN PHẨM GIẢM GIÁ" data={saleProduct} />
       </div>
 
-      {/* Sản phẩm bán chạy */}
+      {/* Dịch vụ thú cưng */}
       <div className="relative mt-[30px] rounded-lg p-6 px-4 sm:px-[40px] lg:px-[154px]">
-        <HotProduct data={hotProduct} />
-      </div>
-
-      {/* Sản phẩm theo danh mục */}
-      {categories.map((category) => (
-        <div
-          key={category._id}
-          className=" mt-[30px] rounded-lg p-6 px-4 sm:px-[40px] lg:px-[154px]"
-        >
-          <div className="mx-auto flex h-[50px] w-full max-w-[900px] items-center justify-center rounded-[40px] bg-[#22A6DF] text-base font-medium text-white md:text-lg">
-            MUA SẮM CHO {category.name.toUpperCase()}
+        <div className="relative ml-[15px] w-[200px] rounded-t-lg border-l border-r border-t border-[#1890ff] px-2 py-2 sm:ml-[30px] sm:w-[250px] sm:px-4 md:w-[300px]">
+          <div className="absolute z-10 px-2 bg-white -top-7 left-3">
+            <img
+              src="/images/icons/paw.png"
+              alt="Paw Icon"
+              className="h-8 w-8 sm:h-12 sm:w-12 md:h-[50px] md:w-[50px]"
+            />
           </div>
-
-          <CateProduct data={productsByCategory[category.name] || []} />
-
-          <div className="mt-6 text-center">
-            <Button className="rounded-md border border-gray-300 px-6 py-5 text-base hover:bg-gray-100">
-              <Link to={`/product?category=${category.name.toLowerCase()}`}>
-                Xem thêm sản phẩm{" "}
-                <span className="font-semibold">dành cho {category.name}</span>
-              </Link>
-            </Button>
-          </div>
+          <h2 className="relative z-20 text-base font-semibold text-center sm:text-lg">
+            DỊCH VỤ THÚ CƯNG
+          </h2>
         </div>
-      ))}
-
-      {/* PetNews */}
-      <div className="w-full bg-white p-3 sm:p-4 md:p-6 lg:p-8 xl:px-[154px]">
-        {/* Brand Logos Section */}
-        <div className="mb-6 grid grid-cols-2 gap-3 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:gap-6">
-          {[
-            { src: "/images/brands/royalcanin.png", alt: "Royal Canin" },
-            { src: "/images/brands/kitcat.png", alt: "Kit Cat" },
-            { src: "/images/brands/gimcat.png", alt: "Gim Cat" },
-            { src: "/images/brands/lapaw.png", alt: "LaPaw" },
-            { src: "/images/brands/tropiclean.png", alt: "TropiClean" },
-          ].map((brand, index) => (
-            <div
-              key={index}
-              className="group flex items-center justify-center p-2 transition-transform duration-300 hover:scale-105"
-            >
+        <div className="mt-4 bg-gradient-to-r from-[#f8e1e1] to-[#e0f7fa] rounded-lg overflow-hidden shadow-lg p-6">
+          <div className="flex flex-col items-center md:flex-row md:items-start">
+            <div className="md:w-1/2">
               <img
-                src={brand.src}
-                alt={brand.alt}
-                className="h-auto max-h-[60px] w-auto object-contain sm:max-h-[80px] md:max-h-[100px]"
-                loading="lazy"
+                src="/images/services/pet-care.jpg"
+                alt="Pet Care Services"
+                className="object-cover w-full h-64 rounded-lg md:rounded-none md:rounded-l-lg"
               />
             </div>
-          ))}
+            <div className="p-4 text-center md:w-1/2 md:p-6 md:text-left">
+              <p className="mb-4 text-sm text-gray-600 sm:text-base md:text-lg">
+                Chúng tôi cung cấp các dịch vụ chăm sóc toàn diện cho thú cưng của bạn, bao gồm tắm rửa, cắt tỉa lông, kiểm tra sức khỏe và tư vấn dinh dưỡng. Hãy để chúng tôi giúp thú cưng của bạn luôn khỏe mạnh và hạnh phúc!
+              </p>
+              <Link to="/services">
+                <button className="px-4 py-2 bg-[#1890ff] text-white rounded-md hover:bg-[#40a9ff] transition-colors text-sm sm:text-base">
+                  Khám phá dịch vụ
+                </button>
+              </Link>
+            </div>
+          </div>
         </div>
+      </div>
 
-        {/* News Section */}
-        <div className="rounded-lg border bg-white p-4 shadow-sm sm:p-6 md:p-8 lg:p-10">
-          {/* News Header */}
-          <div className="mb-4 flex items-center justify-between sm:mb-6">
-            <h3 className="text-sm font-bold uppercase tracking-wider sm:text-base md:text-lg">
+      {/* Bài viết */}
+      <div className="relative mt-[30px] rounded-lg p-6 px-4 sm:px-[40px] lg:px-[154px]">
+        <div className="relative ml-[15px] w-[200px] rounded-t-lg border-l border-r border-t border-[#1890ff] px-2 py-2 sm:ml-[30px] sm:w-[250px] sm:px-4 md:w-[300px]">
+          <div className="absolute z-10 px-2 bg-white -top-7 left-3">
+            <img
+              src="/images/icons/paw.png"
+              alt="Paw Icon"
+              className="h-8 w-8 sm:h-12 sm:w-12 md:h-[50px] md:w-[50px]"
+            />
+          </div>
+          <h2 className="relative z-20 text-base font-semibold text-center sm:text-lg">
+            BÀI VIẾT
+          </h2>
+        </div>
+        <div className="p-4 mt-4 bg-white border rounded-lg shadow-sm sm:p-6 md:p-8 lg:p-10">
+          <div className="flex items-center justify-between mb-4 sm:mb-6">
+            <h3 className="text-sm font-bold tracking-wider uppercase sm:text-base md:text-lg">
               CÓ THỂ BẠN MUỐN BIẾT
             </h3>
             <a
@@ -215,76 +197,60 @@ export default function Home() {
             </a>
           </div>
 
-          {/* Main Article Grid */}
           {blogs.length > 0 ? (
             <div className="grid gap-4 lg:grid-cols-12 lg:gap-6 xl:gap-8">
-              {/* Main Article Image */}
               <div className="lg:col-span-6">
                 <div className="relative h-[200px] w-full overflow-hidden rounded-lg sm:h-[250px] md:h-[300px]">
                   <img
-                    src={blogs[0].image_url || "/images/brands/concho.png"} // Dữ liệu động từ API
+                    src={blogs[0].image_url || "/images/brands/concho.png"}
                     alt={blogs[0].title}
-                    className="h-full w-full object-cover transition-transform duration-300 hover:scale-105"
+                    className="object-cover w-full h-full transition-transform duration-300 hover:scale-105"
                   />
                 </div>
               </div>
-
-              {/* Main Article Content */}
               <div className="flex flex-col lg:col-span-6">
                 <h4 className="mb-2 text-base font-bold leading-tight sm:text-lg md:text-xl">
                   {blogs[0].title}
                 </h4>
-
-                <div className="mb-3 flex flex-wrap gap-3 text-xs text-gray-500 sm:text-sm">
+                <div className="flex flex-wrap gap-3 mb-3 text-xs text-gray-500 sm:text-sm">
                   <span className="flex items-center gap-2">
                     <FaUserEdit className="text-[#22A6DF]" />
                     <span className="flex gap-1">
-                      by{" "}
-                      <span className="font-semibold">{blogs[0].author}</span>
+                      by <span className="font-semibold">{blogs[0].author}</span>
                     </span>
                   </span>
                   <span className="flex items-center gap-2">
                     <FaCalendarAlt className="text-[#22A6DF]" />
-                    <span>
-                      {new Date(blogs[0].createdAt).toLocaleDateString()}
-                    </span>
+                    <span>{new Date(blogs[0].createdAt).toLocaleDateString()}</span>
                   </span>
                 </div>
-
                 <p className="mb-4 text-xs leading-relaxed text-gray-700 sm:text-sm md:mb-6">
-                  {parse(blogs[0].content.slice(0, 1000) + "...")}{" "}
-                  {/* Cắt ngắn nội dung */}
+                  {parse(blogs[0].content.slice(0, 1000) + "...")}
                 </p>
-
                 <Link to={`/blogs/${blogs[0]._id}`}>
                   <button className="group flex items-center gap-2 self-start rounded-md border border-[#22A6DF] px-4 py-2 text-xs font-medium text-[#22A6DF] transition-all hover:bg-[#22A6DF] hover:text-white sm:text-sm">
                     Đọc thêm
-                    <span className="transform transition-transform group-hover:translate-x-1">
-                      »
-                    </span>
+                    <span className="transition-transform transform group-hover:translate-x-1">»</span>
                   </button>
                 </Link>
               </div>
             </div>
           ) : (
-            <p className="text-center text-gray-500">
-              Không có bài viết nào để hiển thị.
-            </p>
+            <p className="text-center text-gray-500">Không có bài viết nào để hiển thị.</p>
           )}
 
-          {/* Related Articles */}
-          <div className="mt-6 grid gap-4 sm:mt-8 sm:grid-cols-2 sm:gap-6 lg:grid-cols-3">
+          <div className="grid gap-4 mt-6 sm:mt-8 sm:grid-cols-2 sm:gap-6 lg:grid-cols-3">
             {blogs.slice(1, 4).map((blog, index) => (
               <Link
                 to={`/blogs/${blog._id}`}
                 key={index}
-                className="group flex items-start gap-3 sm:gap-4"
+                className="flex items-start gap-3 group sm:gap-4"
               >
                 <div className="relative h-[80px] w-[100px] min-w-[100px] overflow-hidden rounded-lg sm:h-[100px] sm:w-[120px] sm:min-w-[120px] md:h-[120px] md:w-[140px] md:min-w-[140px]">
                   <img
-                    src={blog.image_url || "/images/brands/concho.png"} // Dữ liệu động từ API
+                    src={blog.image_url || "/images/brands/concho.png"}
                     alt={blog.title}
-                    className="h-full w-full object-cover transition-transform duration-300 group-hover:scale-105"
+                    className="object-cover w-full h-full transition-transform duration-300 group-hover:scale-105"
                   />
                 </div>
                 <div className="flex flex-col">
