+++ FE/tailwind.config.js
@@ -5,6 +5,15 @@ export const content = [
   "./components/**/*.{js,ts,jsx,tsx}",
 ];
 export const theme = {
-  extend: {},
+  extend: {
+     colors: {
+        primary: {
+          300: '#7DCAEB',
+          400: '#4FB8E5',
+          500: '#22A6DF',
+          600: '#1E90C2', 
+        },
+      },
+  },
 };
 export const plugins = [];
