+++ layout.html
@@ -1,0 +1,110 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Sản phẩm bán chạy</title>
+  <script src="https://cdn.tailwindcss.com"></script>
+</head>
+
+<body class="p-6 bg-gray-100">
+  <!-- Title -->
+  <div class="flex items-center w-full bg-red-100 rounded-md">
+    <span class="px-4 py-2 text-sm font-bold text-white bg-red-500 rounded-l-md">
+      SẢN PHẨM BÁN CHẠY
+    </span>
+  </div>
+
+  <!-- Products Grid -->
+  <div class="grid grid-cols-1 gap-6 py-6 sm:grid-cols-2 lg:grid-cols-4">
+    <!-- Product Card -->
+    <div class="w-full p-4 transition bg-white border border-gray-200 rounded-lg shadow hover:shadow-lg">
+      <!-- Image -->
+      <div class="h-[230px] flex items-center justify-center">
+        <img src="https://product.hstatic.net/1000356051/product/combo_36_goi_85g_-_pate_cho_royal_canin_mini_adult_gravy_petmall_e246fc4338f541d989164125aa0c61d9_master.png"
+          alt="Royal Canin" class="object-cover w-full h-full rounded-md" />
+      </div>
+      <!-- Price + Wishlist -->
+      <div class="flex items-center justify-between mt-3">
+        <span class="text-lg font-bold text-red-600">1.374.000đ</span>
+        <button class="text-gray-500 hover:text-red-500">❤️</button>
+      </div>
+      <!-- Name -->
+      <p class="mt-2 text-sm text-gray-700 line-clamp-2">
+        Combo 36 Gói 85G - Pate chó Royal Canin Mini Adult Gravy Petmall
+      </p>
+      <!-- Button -->
+      <div class="flex justify-center">
+        <button
+          class="px-6 py-2 mt-3 font-medium text-white transition bg-red-500 rounded-md hover:bg-red-600">
+          Mua ngay
+        </button>
+      </div>
+    </div>
+
+    <!-- Copy thêm 3 sản phẩm -->
+    <div class="w-full p-4 transition bg-white border border-gray-200 rounded-lg shadow hover:shadow-lg">
+      <div class="h-[230px] flex items-center justify-center">
+        <img src="https://down-vn.img.susercontent.com/file/vn-11134207-7r98o-lphfh9b4znaz53"
+          alt="Dog Food" class="object-cover w-full h-full rounded-md" />
+      </div>
+      <div class="flex items-center justify-between mt-3">
+        <span class="text-lg font-bold text-red-600">890.000đ</span>
+        <button class="text-gray-500 hover:text-red-500">❤️</button>
+      </div>
+      <p class="mt-2 text-sm text-gray-700 line-clamp-2">
+        Thức ăn hạt cho chó con Pedigree 20kg
+      </p>
+      <div class="flex justify-center">
+        <button
+          class="px-6 py-2 mt-3 font-medium text-white transition bg-red-500 rounded-md hover:bg-red-600">
+          Mua ngay
+        </button>
+      </div>
+    </div>
+
+    <div class="w-full p-4 transition bg-white border border-gray-200 rounded-lg shadow hover:shadow-lg">
+      <div class="h-[230px] flex items-center justify-center">
+        <img src="https://bizweb.dktcdn.net/thumb/large/100/369/010/products/screenshot-51.png"
+          alt="Whiskas" class="object-cover w-full h-full rounded-md" />
+      </div>
+      <div class="flex items-center justify-between mt-3">
+        <span class="text-lg font-bold text-red-600">560.000đ</span>
+        <button class="text-gray-500 hover:text-red-500">❤️</button>
+      </div>
+      <p class="mt-2 text-sm text-gray-700 line-clamp-2">
+        Thức ăn Whiskas cho mèo trưởng thành 20kg
+      </p>
+      <div class="flex justify-center">
+        <button
+          class="px-6 py-2 mt-3 font-medium text-white transition bg-red-500 rounded-md hover:bg-red-600">
+          Mua ngay
+        </button>
+      </div>
+    </div>
+
+    <div class="w-full p-4 transition bg-white border border-gray-200 rounded-lg shadow hover:shadow-lg">
+      <div class="h-[230px] flex items-center justify-center">
+        <img src="https://cf.shopee.vn/file/72cabb5344ff9f3f839b272b8dc76eb3"
+          alt="Cat Food" class="object-cover w-full h-full rounded-md" />
+      </div>
+      <div class="flex items-center justify-between mt-3">
+        <span class="text-lg font-bold text-red-600">1.200.000đ</span>
+        <button class="text-gray-500 hover:text-red-500">❤️</button>
+      </div>
+      <p class="mt-2 text-sm text-gray-700 line-clamp-2">
+        Combo 24 lon thức ăn cho mèo Me-O Tuna 400g
+      </p>
+      <div class="flex justify-center">
+        <button
+          class="px-6 py-2 mt-3 font-medium text-white transition bg-red-500 rounded-md hover:bg-red-600">
+          Mua ngay
+        </button>
+      </div>
+    </div>
+
+  </div>
+</body>
+
+</html>
